Misc styling updates

+++ resources/views/livewire/notice/dashboard/latest-notice-list.blade.php
@@ -9,7 +9,7 @@
             </th>
           </tr>
           @endif
-          @if (count($notices) > 0)
+          @if (false && count($notices) > 0)
             @foreach ($notices as $notice) 
             <tr class="border-bottom">
               <td class="text-primary-rm font-weight-bold-rm">
@@ -29,7 +29,10 @@
           @else
             <tr>
               <td class="text-muted">
-                No notice to show.
+                 <div class="h5 font-weight-bold px-3" style="color: orange;">
+                   <i class="fas fa-exclamation-circle mr-1"></i>
+                   No notice.
+                 </div>
               </td>
             </tr>
           @endif
